Implemented BaseGridWidget class to remove duplicated code
